+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.0.2 - 2022-09-30
+ - Most of the important TODO items implemented (dynamic check/updates to presets, defaults) Removed the need for static
+   .conf file with defaults. Still need to implement dynamic nested_list generation (interface in html)
 ## 0.0.1 - 2022-08-04
  - A prototype released for testing by other members of the group. Basic functionality, some bugs may be still present.